WIP of Porting NetworkController to Ts (does not compile at this point)
