--- packages/cli/src/models/network/index.js
@@ -1,7 +1,0 @@
-import ControllerFor from './ControllerFor'
-import NetworkController from './NetworkController'
-
-export default {
-  ControllerFor,
-  NetworkController,
-}
