+++ packages/cli/src/models/network/index.ts
@@ -1,0 +1,7 @@
+import ControllerFor from './ControllerFor';
+import NetworkController from './NetworkController';
+
+export default {
+  ControllerFor,
+  NetworkController,
+};
